+++ config/role.go
@@ -1,7 +1,7 @@
 package config
 
 type Role struct {
-	ID    int    `gorm:"primary_key;AUTO_INCREMENT" yaml:"id" json:"id"`
+	ID    uint64 `gorm:"primary_key;AUTO_INCREMENT" yaml:"id" json:"id"`
 	Name  string `yaml:"name" json:"name"`
 	Users []User `yaml:"users" json:"users"`
 	//Users      []User `gorm:"many2many:permissions"`
@@ -12,11 +12,11 @@ func NewRole() *Role {
 	return &Role{}
 }
 
-func (role *Role) SetID(id int) {
+func (role *Role) SetID(id uint64) {
 	role.ID = id
 }
 
-func (role *Role) GetID() int {
+func (role *Role) GetID() uint64 {
 	return role.ID
 }
